Enable browser caching and improve performance.
Caching improvements:
- Chunked uploading to database
- Ignore unnecessary locale files

Additionally, some code cleanup.

+++ bci/util.py
@@ -7,6 +7,7 @@ import logging
 import os
 import shutil
 import time
+from typing import Optional
 
 import requests
 
@@ -42,11 +43,13 @@ def copy_folder(src_path, dst_path):
     shutil.copytree(src_path, dst_path, dirs_exist_ok=True)
 
 
-def remove_all_in_folder(folder_path: str) -> None:
+def remove_all_in_folder(folder_path: str, except_files: Optional[list[str]]=None) -> None:
+    except_files = [] if except_files is None else except_files
     for root, dirs, files in os.walk(folder_path):
         for file_name in files:
             file_path = os.path.join(root, file_name)
-            os.remove(file_path)
+            if file_name not in except_files:
+                os.remove(file_path)
         for dir_name in dirs:
             dir_path = os.path.join(root, dir_name)
             shutil.rmtree(dir_path)
